+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for HELIX Component Loader interface.
+
+### Fixed
+- Blind HELIX Components now correctly conform to the Component interface.
 
 ## [0.1.0] - 2021-09-24
 ### Added